+++ lib/ui/widget/alice_call_list_item_widget.dart
@@ -90,13 +90,22 @@ class AliceCallListItemWidget extends StatelessWidget {
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
-        Flexible(flex:1, child: Text(_formatTime(call.request.time), style: TextStyle(fontSize: 12))),
-    Flexible(flex:1,child:Text("${AliceConversionHelper.formatTime(call.duration)}",
-            style: TextStyle(fontSize: 12))),
-    Flexible(flex:1,child:Text(
+        Flexible(
+            flex: 1,
+            child: Text(_formatTime(call.request.time),
+                style: TextStyle(fontSize: 12))),
+        Flexible(
+            flex: 1,
+            child: Text("${AliceConversionHelper.formatTime(call.duration)}",
+                style: TextStyle(fontSize: 12))),
+        Flexible(
+          flex: 1,
+          child: Text(
             "${AliceConversionHelper.formatBytes(call.request.size)} / "
             "${AliceConversionHelper.formatBytes(call.response.size)}",
-            style: TextStyle(fontSize: 12),))
+            style: TextStyle(fontSize: 12),
+          ),
+        )
       ],
     );
   }
@@ -162,7 +171,7 @@ class AliceCallListItemWidget extends StatelessWidget {
     if (status == -1) {
       return AliceConstants.red;
     } else if (status < 200) {
-      return Theme.of(context).textTheme.body1.color;
+      return Theme.of(context).textTheme.bodyText1.color;
     } else if (status >= 200 && status < 300) {
       return AliceConstants.green;
     } else if (status >= 300 && status < 400) {
@@ -170,7 +179,7 @@ class AliceCallListItemWidget extends StatelessWidget {
     } else if (status >= 400 && status < 600) {
       return AliceConstants.red;
     } else {
-      return Theme.of(context).textTheme.body1.color;
+      return Theme.of(context).textTheme.bodyText1.color;
     }
   }